begin to clean up package
