+++ include/peak_cam/acquisition_parameters.hpp
@@ -33,78 +33,34 @@
 // CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE 
 // POSSIBILITY OF SUCH DAMAGE. 
+#ifndef PEAK_CAM__ACQUISITION_PARAMETERS_HPP_
+#define PEAK_CAM__ACQUISITION_PARAMETERS_HPP_
 
 
-
-#pragma once
-
-#include <iostream>
-#include <atomic>
-
-//ROS Headers
-#include <ros/ros.h>
-#include <std_msgs/Float64.h>
-#include <sensor_msgs/Image.h>
-#include <sensor_msgs/image_encodings.h>
-#include <dynamic_reconfigure/server.h>
-
-//OpenCV Headers
-#include <opencv2/opencv.hpp>
-#include <cv_bridge/cv_bridge.h>
-#include "opencv2/imgproc/imgproc.hpp"
-#include "opencv2/highgui/highgui.hpp"
-#include <cv_bridge/cv_bridge.h>
-
-//IDS Camera Headers
 #include <peak_ipl/peak_ipl.hpp>
-#include <peak/converters/peak_buffer_converter_ipl.hpp>
-#include <peak/peak.hpp>
-
-//Parameters
-#include "acquisition_parameters.hpp"
-
-#include <peak_cam/PeakCamConfig.h>
+#include <sensor_msgs/image_encodings.h>
+#include <sensor_msgs/Image.h>
 
 
 namespace peak_cam
 {
 
-class Peak_Cam
+struct Peak_Params
 {
-    using Config = PeakCamConfig;
-    using ReconfigureServer = dynamic_reconfigure::Server<Config>;
-
-public:
-    Peak_Cam(ros::NodeHandle nh);
-    ~Peak_Cam();
-    
-    // acquisitionLoop function and bool are public to run on particular thread
-    void acquisitionLoop();
-    
-    // Preventing two threads to acces variable acquisitionLoop_running
-    std::atomic<bool> acquisitionLoop_running{false};
-
-private:
-    ros::NodeHandle nh_private;
-
-    void reconfigureRequest(const Config &, uint32_t);
-    void openDevice();
-    void setDeviceParameters();
-    void closeDevice();
-
-    ros::Publisher image_publisher;
-
-    dynamic_reconfigure::Server<Config> server;
-    dynamic_reconfigure::Server<Config>::CallbackType f;
-
-    std::shared_ptr<peak::core::DataStream> m_dataStream;
-    std::shared_ptr<peak::core::Device> m_device;
-    std::shared_ptr<peak::core::NodeMap> m_nodeMapRemoteDevice;
-    peak::ipl::PixelFormatName pixel_format_name;
-    sensor_msgs::Image image_for_encoding;
-
-    // Camera Parameters
-    Peak_Params peak_params;
+    std::string selectedDevice{"000000"}; // default to all 0's
+    int ExposureTime{100};
+    int AcquisitionFrameRate{1};
+    int ImageHeight{480};
+    int ImageWidth{640};
+    double Gamma{1.2};
+    std::string ExposureAuto{"Off"};
+    std::string GainAuto{"Off"};
+    std::string GainSelector;
+    std::string PixelFormat;
+    std::string TriggerMode{"Off"};
+    int TriggerSource{0};
 };
 
-} // namespace peak_cam
+}
+
+#endif  // PEAK_CAM__ACQUISITION_PARAMETERS_HPP_
